Added cp to notifications
For shinies that are common, whether it's an event that is spawning lots of the same pokemon or a nest, I don't like to catch them if they aren't a high level. So I added cp to the discord notification to help me decide if I want to log in and catch that shiny.